+++ headwater-bitmap/src/main/java/headwater/bitmap/BitmapChunker.java
@@ -13,7 +13,7 @@ public class BitmapChunker {
      * @return
      */
     public static List<Chunk> divide(IBitmap bm, int chunkArraySize, boolean skipEmpties) {
-        final int numChunks = (bm.getBitLength() / (chunkArraySize *8)) + ((bm.getBitLength() % (chunkArraySize * 8)) > 0 ? 1 : 0);
+        final int numChunks = (int)((bm.getBitLength() / (chunkArraySize *8)) + ((bm.getBitLength() % (chunkArraySize * 8)) > 0 ? 1 : 0));
         List<Chunk> chunks = new ArrayList<Chunk>(numChunks);
         
         for (int i = 0; i < numChunks; i++) {
